Issue / Data Table Export

+++ test/recipe/Baked.Test.Recipe.Service/Theme/TableWithFooter.cs
@@ -1,3 +1,3 @@
 ﻿namespace Baked.Test.Theme;
 
-public record TableWithFooter(IEnumerable<TableRow> Items, int FooterColumn1, int FooterColumn2);
+public record TableWithFooter(IEnumerable<TableRow> Items, int FooterColumn1, double FooterColumn2);
